Replaced flask based v1. Not currently passing integration tests, need to change them to work with new implementation
