+++ app/router/__init__.py
@@ -1,8 +1,5 @@
 from fastapi import APIRouter
 
-# Create the router.
-# router = APIRouter()
-
 # The routes.
 from .v2 import latest, sources, locations  # isort:skip
-# from .v1 import confirmed, deaths, recovered, all
+from .v1 import confirmed, deaths, recovered, all
