+++ app/router/v1/__init__.py
@@ -1,0 +1,3 @@
+from fastapi import APIRouter
+
+router = APIRouter()
